Separated structures from the main v2.go code
