+++ internal/api/routes.go
@@ -53,7 +53,9 @@ func (a *API) InitRoutes() {
 			AuthenticationMW(a.oa2Prov, a.log),
 			ValidateComponentsMW(a.db, a.log),
 			v2.PostIncidentExtractHandler(a.db, a.log))
-
+		// v2API.GET(("incidents/:id/updates"), v2.GetEventUpdatesHandler(a.db, a.log))
+		// v2API.PATCH("incidents/:id/updates/:update_id",
+		// AuthenticationMW(a.oa2Prov, a.log), v2.PatchEventUpdateHandler(a.db, a.log))
 		v2API.GET("availability", v2.GetComponentsAvailabilityHandler(a.db, a.log))
 
 		// For testing purposes only
